requirements.txt
numpy>=1.23.0
Pillow>=9.2.0
